Remove typescript source
