+++ .gitignore
@@ -1,3 +1,2 @@
 coverage/
-lib
 node_modules/
