+++ docs/assets/highlight.css
@@ -27,8 +27,6 @@
     --dark-hl-12: #D7BA7D;
     --light-hl-13: #098658;
     --dark-hl-13: #B5CEA8;
-    --light-hl-14: #267F99;
-    --dark-hl-14: #4EC9B0;
     --light-code-background: #FFFFFF;
     --dark-code-background: #1E1E1E;
 }
@@ -48,7 +46,6 @@
     --hl-11: var(--light-hl-11);
     --hl-12: var(--light-hl-12);
     --hl-13: var(--light-hl-13);
-    --hl-14: var(--light-hl-14);
     --code-background: var(--light-code-background);
 } }
 
@@ -67,7 +64,6 @@
     --hl-11: var(--dark-hl-11);
     --hl-12: var(--dark-hl-12);
     --hl-13: var(--dark-hl-13);
-    --hl-14: var(--dark-hl-14);
     --code-background: var(--dark-code-background);
 } }
 
@@ -86,7 +82,6 @@
     --hl-11: var(--light-hl-11);
     --hl-12: var(--light-hl-12);
     --hl-13: var(--light-hl-13);
-    --hl-14: var(--light-hl-14);
     --code-background: var(--light-code-background);
 }
 
@@ -105,7 +100,6 @@
     --hl-11: var(--dark-hl-11);
     --hl-12: var(--dark-hl-12);
     --hl-13: var(--dark-hl-13);
-    --hl-14: var(--dark-hl-14);
     --code-background: var(--dark-code-background);
 }
 
@@ -123,5 +117,4 @@
 .hl-11 { color: var(--hl-11); }
 .hl-12 { color: var(--hl-12); }
 .hl-13 { color: var(--hl-13); }
-.hl-14 { color: var(--hl-14); }
 pre, code { background: var(--code-background); }
