+++ docs/assets/search.js
@@ -1,1 +1,1 @@
-window.searchData = JSON.parse("{\"rows\":[{\"kind\":256,\"name\":\"WrapperOptions\",\"url\":\"interfaces/WrapperOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"indent\",\"url\":\"interfaces/WrapperOptions.html#indent\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"indentChar\",\"url\":\"interfaces/WrapperOptions.html#indentChar\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"indentEmpty\",\"url\":\"interfaces/WrapperOptions.html#indentEmpty\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"indentFirst\",\"url\":\"interfaces/WrapperOptions.html#indentFirst\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"isEmpty\",\"url\":\"interfaces/WrapperOptions.html#isEmpty\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"isNewline\",\"url\":\"interfaces/WrapperOptions.html#isNewline\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"locale\",\"url\":\"interfaces/WrapperOptions.html#locale\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"newline\",\"url\":\"interfaces/WrapperOptions.html#newline\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"newlineReplacement\",\"url\":\"interfaces/WrapperOptions.html#newlineReplacement\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"trim\",\"url\":\"interfaces/WrapperOptions.html#trim\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/WrapperOptions.html#width\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":2048,\"name\":\"escape\",\"url\":\"interfaces/WrapperOptions.html#escape\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":128,\"name\":\"SegmentWrapper\",\"url\":\"classes/SegmentWrapper.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/SegmentWrapper.html#constructor\",\"classes\":\"\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#opts\",\"url\":\"classes/SegmentWrapper.html#_opts\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#words\",\"url\":\"classes/SegmentWrapper.html#_words\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#graphemes\",\"url\":\"classes/SegmentWrapper.html#_graphemes\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#indent\",\"url\":\"classes/SegmentWrapper.html#_indent\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#width\",\"url\":\"classes/SegmentWrapper.html#_width\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":2048,\"name\":\"wrap\",\"url\":\"classes/SegmentWrapper.html#wrap\",\"classes\":\"\",\"parent\":\"SegmentWrapper\"},{\"kind\":2048,\"name\":\"#graphemeCount\",\"url\":\"classes/SegmentWrapper.html#_graphemeCount\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,27.3]],[\"comment/0\",[]],[\"name/1\",[1,22.192]],[\"comment/1\",[]],[\"name/2\",[2,27.3]],[\"comment/2\",[]],[\"name/3\",[3,27.3]],[\"comment/3\",[]],[\"name/4\",[4,27.3]],[\"comment/4\",[]],[\"name/5\",[5,27.3]],[\"comment/5\",[]],[\"name/6\",[6,27.3]],[\"comment/6\",[]],[\"name/7\",[7,27.3]],[\"comment/7\",[]],[\"name/8\",[8,27.3]],[\"comment/8\",[]],[\"name/9\",[9,27.3]],[\"comment/9\",[]],[\"name/10\",[10,27.3]],[\"comment/10\",[]],[\"name/11\",[11,22.192]],[\"comment/11\",[]],[\"name/12\",[12,27.3]],[\"comment/12\",[]],[\"name/13\",[13,27.3]],[\"comment/13\",[]],[\"name/14\",[14,27.3]],[\"comment/14\",[]],[\"name/15\",[15,27.3]],[\"comment/15\",[]],[\"name/16\",[16,27.3]],[\"comment/16\",[]],[\"name/17\",[17,27.3]],[\"comment/17\",[]],[\"name/18\",[1,22.192]],[\"comment/18\",[]],[\"name/19\",[11,22.192]],[\"comment/19\",[]],[\"name/20\",[18,27.3]],[\"comment/20\",[]],[\"name/21\",[19,27.3]],[\"comment/21\",[]]],\"invertedIndex\":[[\"constructor\",{\"_index\":14,\"name\":{\"14\":{}},\"comment\":{}}],[\"escape\",{\"_index\":12,\"name\":{\"12\":{}},\"comment\":{}}],[\"graphemecount\",{\"_index\":19,\"name\":{\"21\":{}},\"comment\":{}}],[\"graphemes\",{\"_index\":17,\"name\":{\"17\":{}},\"comment\":{}}],[\"indent\",{\"_index\":1,\"name\":{\"1\":{},\"18\":{}},\"comment\":{}}],[\"indentchar\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"indentempty\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"indentfirst\",{\"_index\":4,\"name\":{\"4\":{}},\"comment\":{}}],[\"isempty\",{\"_index\":5,\"name\":{\"5\":{}},\"comment\":{}}],[\"isnewline\",{\"_index\":6,\"name\":{\"6\":{}},\"comment\":{}}],[\"locale\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"newline\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"newlinereplacement\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"opts\",{\"_index\":15,\"name\":{\"15\":{}},\"comment\":{}}],[\"segmentwrapper\",{\"_index\":13,\"name\":{\"13\":{}},\"comment\":{}}],[\"trim\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"width\",{\"_index\":11,\"name\":{\"11\":{},\"19\":{}},\"comment\":{}}],[\"words\",{\"_index\":16,\"name\":{\"16\":{}},\"comment\":{}}],[\"wrap\",{\"_index\":18,\"name\":{\"20\":{}},\"comment\":{}}],[\"wrapperoptions\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
+window.searchData = JSON.parse("{\"rows\":[{\"kind\":128,\"name\":\"SegmentWrapper\",\"url\":\"classes/SegmentWrapper.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/SegmentWrapper.html#constructor\",\"classes\":\"\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#opts\",\"url\":\"classes/SegmentWrapper.html#_opts\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#words\",\"url\":\"classes/SegmentWrapper.html#_words\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#graphemes\",\"url\":\"classes/SegmentWrapper.html#_graphemes\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#indent\",\"url\":\"classes/SegmentWrapper.html#_indent\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":1024,\"name\":\"#width\",\"url\":\"classes/SegmentWrapper.html#_width\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":2048,\"name\":\"wrap\",\"url\":\"classes/SegmentWrapper.html#wrap\",\"classes\":\"\",\"parent\":\"SegmentWrapper\"},{\"kind\":2048,\"name\":\"#graphemeCount\",\"url\":\"classes/SegmentWrapper.html#_graphemeCount\",\"classes\":\"tsd-is-private\",\"parent\":\"SegmentWrapper\"},{\"kind\":4194304,\"name\":\"EscapeString\",\"url\":\"types/EscapeString.html\",\"classes\":\"\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/EscapeString.html#__type\",\"classes\":\"\",\"parent\":\"EscapeString\"},{\"kind\":256,\"name\":\"WrapperOptions\",\"url\":\"interfaces/WrapperOptions.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"indent\",\"url\":\"interfaces/WrapperOptions.html#indent\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"indentChar\",\"url\":\"interfaces/WrapperOptions.html#indentChar\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"indentEmpty\",\"url\":\"interfaces/WrapperOptions.html#indentEmpty\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"indentFirst\",\"url\":\"interfaces/WrapperOptions.html#indentFirst\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"isEmpty\",\"url\":\"interfaces/WrapperOptions.html#isEmpty\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"isNewline\",\"url\":\"interfaces/WrapperOptions.html#isNewline\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"locale\",\"url\":\"interfaces/WrapperOptions.html#locale\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"newline\",\"url\":\"interfaces/WrapperOptions.html#newline\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"newlineReplacement\",\"url\":\"interfaces/WrapperOptions.html#newlineReplacement\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"trim\",\"url\":\"interfaces/WrapperOptions.html#trim\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"width\",\"url\":\"interfaces/WrapperOptions.html#width\",\"classes\":\"\",\"parent\":\"WrapperOptions\"},{\"kind\":1024,\"name\":\"escape\",\"url\":\"interfaces/WrapperOptions.html#escape\",\"classes\":\"\",\"parent\":\"WrapperOptions\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,28.134]],[\"comment/0\",[]],[\"name/1\",[1,28.134]],[\"comment/1\",[]],[\"name/2\",[2,28.134]],[\"comment/2\",[]],[\"name/3\",[3,28.134]],[\"comment/3\",[]],[\"name/4\",[4,28.134]],[\"comment/4\",[]],[\"name/5\",[5,23.026]],[\"comment/5\",[]],[\"name/6\",[6,23.026]],[\"comment/6\",[]],[\"name/7\",[7,28.134]],[\"comment/7\",[]],[\"name/8\",[8,28.134]],[\"comment/8\",[]],[\"name/9\",[9,28.134]],[\"comment/9\",[]],[\"name/10\",[10,28.134]],[\"comment/10\",[]],[\"name/11\",[11,28.134]],[\"comment/11\",[]],[\"name/12\",[5,23.026]],[\"comment/12\",[]],[\"name/13\",[12,28.134]],[\"comment/13\",[]],[\"name/14\",[13,28.134]],[\"comment/14\",[]],[\"name/15\",[14,28.134]],[\"comment/15\",[]],[\"name/16\",[15,28.134]],[\"comment/16\",[]],[\"name/17\",[16,28.134]],[\"comment/17\",[]],[\"name/18\",[17,28.134]],[\"comment/18\",[]],[\"name/19\",[18,28.134]],[\"comment/19\",[]],[\"name/20\",[19,28.134]],[\"comment/20\",[]],[\"name/21\",[20,28.134]],[\"comment/21\",[]],[\"name/22\",[6,23.026]],[\"comment/22\",[]],[\"name/23\",[21,28.134]],[\"comment/23\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":10,\"name\":{\"10\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":1,\"name\":{\"1\":{}},\"comment\":{}}],[\"escape\",{\"_index\":21,\"name\":{\"23\":{}},\"comment\":{}}],[\"escapestring\",{\"_index\":9,\"name\":{\"9\":{}},\"comment\":{}}],[\"graphemecount\",{\"_index\":8,\"name\":{\"8\":{}},\"comment\":{}}],[\"graphemes\",{\"_index\":4,\"name\":{\"4\":{}},\"comment\":{}}],[\"indent\",{\"_index\":5,\"name\":{\"5\":{},\"12\":{}},\"comment\":{}}],[\"indentchar\",{\"_index\":12,\"name\":{\"13\":{}},\"comment\":{}}],[\"indentempty\",{\"_index\":13,\"name\":{\"14\":{}},\"comment\":{}}],[\"indentfirst\",{\"_index\":14,\"name\":{\"15\":{}},\"comment\":{}}],[\"isempty\",{\"_index\":15,\"name\":{\"16\":{}},\"comment\":{}}],[\"isnewline\",{\"_index\":16,\"name\":{\"17\":{}},\"comment\":{}}],[\"locale\",{\"_index\":17,\"name\":{\"18\":{}},\"comment\":{}}],[\"newline\",{\"_index\":18,\"name\":{\"19\":{}},\"comment\":{}}],[\"newlinereplacement\",{\"_index\":19,\"name\":{\"20\":{}},\"comment\":{}}],[\"opts\",{\"_index\":2,\"name\":{\"2\":{}},\"comment\":{}}],[\"segmentwrapper\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"trim\",{\"_index\":20,\"name\":{\"21\":{}},\"comment\":{}}],[\"width\",{\"_index\":6,\"name\":{\"6\":{},\"22\":{}},\"comment\":{}}],[\"words\",{\"_index\":3,\"name\":{\"3\":{}},\"comment\":{}}],[\"wrap\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"wrapperoptions\",{\"_index\":11,\"name\":{\"11\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
